+++ klum-ast-core/src/main/java/com/blackbuild/groovy/configdsl/transform/DSL.java
@@ -23,9 +23,7 @@
  */
 package com.blackbuild.groovy.configdsl.transform;
 
-import com.blackbuild.groovy.configdsl.transform.ast.DSLASTTransformation;
-import com.blackbuild.groovy.configdsl.transform.ast.DelegatesToRWTransformation;
-import com.blackbuild.groovy.configdsl.transform.ast.mutators.ModelVerifierTransformation;
+import com.blackbuild.groovy.configdsl.transform.ast.modules.ModuleConfigurator;
 import org.codehaus.groovy.transform.GroovyASTTransformationClass;
 
 import java.lang.annotation.Documented;
@@ -425,7 +423,7 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.CLASS)
 @Inherited // This is currently not used, see https://issues.apache.org/jira/browse/GROOVY-6765
-@GroovyASTTransformationClass(classes={DSLASTTransformation.class, ModelVerifierTransformation.class, DelegatesToRWTransformation.class})
+@GroovyASTTransformationClass(classes= ModuleConfigurator.class)
 @Documented
 public @interface DSL {
     /**