+++ klum-ast-core/src/main/java/com/blackbuild/groovy/configdsl/transform/ast/modules/KlumAstTransformation.java
@@ -1,0 +1,58 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015-2017 Stephan Pauxberger
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.blackbuild.groovy.configdsl.transform.ast.modules;
+
+import com.blackbuild.groovy.configdsl.transform.DSL;
+import org.codehaus.groovy.ast.ASTNode;
+import org.codehaus.groovy.ast.AnnotationNode;
+import org.codehaus.groovy.ast.ClassNode;
+import org.codehaus.groovy.control.SourceUnit;
+import org.codehaus.groovy.transform.AbstractASTTransformation;
+
+import static org.codehaus.groovy.ast.ClassHelper.make;
+
+public abstract class KlumAstTransformation extends AbstractASTTransformation {
+
+    public static final ClassNode DSL_CONFIG_ANNOTATION = make(DSL.class);
+
+    public ClassNode annotatedClass;
+    protected AnnotationNode dslAnnotation;
+
+
+    @Override
+    public final void visit(ASTNode[] nodes, SourceUnit source) {
+        init(nodes, source);
+
+        annotatedClass = (ClassNode) nodes[1];
+        dslAnnotation = (AnnotationNode) nodes[0];
+
+        if (!hasAnnotation(annotatedClass, DSL_CONFIG_ANNOTATION))
+            return;
+
+
+        doVisit();
+    }
+
+    protected abstract void doVisit();
+}